requirements.txt
attrs==18.1.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.2
django-debug-toolbar==1.8
django-extensions==1.7.9
django-filter==1.1.0
djangorestframework==3.6.3
factory-boy==2.11.1
Faker==0.8.15
gunicorn==19.7.1
idna==2.6
itypes==1.1.0
Jinja2==2.10
jsondiff==1.1.2
MarkupSafe==1.0
pluggy==0.6.0
psycopg2==2.7.1
py==1.5.3
pytest==3.4.2
pytest-django==3.1.2
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
six==1.11.0
sqlparse==0.2.4
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.22
